Some compiling warnings removed. Basically, unsigned int -> size_t

+++ RLSimion-Lib/DQN.h
@@ -25,8 +25,8 @@ protected:
 	CFeatureList* m_pStateOutFeatures;
 	CSingleDimensionDiscreteActionVariableGrid* m_pGrid;
 
-	int m_numberOfActions;
-	int m_numberOfStateFeatures;
+	size_t m_numberOfActions;
+	size_t m_numberOfStateFeatures;
 	std::vector<float> m_stateVector;
 	std::vector<float> m_actionValuePredictionVector;
